WIP Centralized connect and functions to client.rs
This is a WIP, but functional just before addressing the addition of the replier functionality.
